services/feed-ingestion/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
celery==5.3.4
celery[sqlalchemy]==5.3.4
requests==2.31.0
python-dotenv==1.0.0
feedparser==6.0.10
pydantic==2.5.0
python-multipart==0.0.6 